requirements.txt
streamlit
llama-parse
langchain-text-splitters
python-dotenv
jinaai
pinecone-client
google-generativeai
